feat: add eth_accounts
